Update conf file format
Update the .conf files to use a very simple format, with lines in any
order of the form

key=value

Add major, minor, and patch versions (controlled by version.py) to
the conf files.

Issue a warning if the major.minor version is outdated. Don't accept
conf files from before this (that is, those without version
information).

Add an extra configuration field for additional flags to provide to the
batch lauch command.

The direct motivation for the new field is to allow usage on the new Piz Daint, which
requires "sbatch -C gpu" to run on the "GPU" portion (previously "Piz
Daint") and "sbatch -C mc" to run on the "multicore" portion (previously
"Piz Dora")