Harness, launcher: add logic for --purge_output
In the __init__() function for the harness, we call the clean()
function, unless we are only verifying. That is, if
we are going to execute the tests, we should delete any existing output.

If --purge_output (or -p) is supplied, the code simply exits at this point.
Verifying and then purging is not supported (but this is not a very useful
combination).

There are two TODOs remaining in launch.py, related to risks associated with
deleting files. These should be discussed and hopefully improved before merging this branch.

1. If in sandbox mode, we delete and entire subtree to wipe out the sandbox
directories. This seems dangerous if somehow the wrong argument crept in for the sandbox_path
2. We delete any custom comparison_file's . This is dangerous if the user supplies the wrong file and has it deleted.