Tests: update to work with new --sandbox and --purge_output functionality
We specify several paths as absolute paths (by using a helper function
to convert paths relative to the python source).

In example 2, we create the output/ directory if it doesn't exist. Note
that this means that the test output and the custom output file 
end up in different places when using sandbox mode.