Rename pthUnitTest to Test
Remove all mention of "unit", since this framework can be used for more general types of testing.

Remove the prefix "pth" from the pthUnitTest class as well.

Many associated cleanups and edits.

An alias is left to pthUnitTest as an alias for the new Test class,
to avoid breaking old code. Use of this should be considered deprecated,
but we don't currently issue any warning about this.