Harness/Launcher: adjust execution output
Reduce redundancy by not having both the launcher and the harness output that they are executing.

Put the command executed on its own line, for ease of copy-pasting.

If present, but the note about the sandbox directory on the same line as the note about which test is being executed.

For multiple-command tests, when not using a batch system, run each
command after an output line, and display the number of the command.