Relative floating point test: add epsilon offset for denominator
This replaces the previous, dubious practice of ignoring comparison with
zero values. Rather, the user can now specify an option "epsilon" to add
to the denominator in relative tests. This is practically useful in
situations where two "numerical zeros" (the magnitude of which is
application-dependent) have a large relative difference.

Update example 9.