Job: remove DAG impementation
Job needs some refactoring, and it's not totally clear that there is
a large practical difference between the DAG implementation and the
Sequence implementation. This is especially true given the current
way we launch jobs, in a single batch file which doesn't allow for the
parallelism which might make a DAG attrative in some scenario.

Overall, the simplification to the codebase seems worth removing this.