Support Python 2.7
This makes a number of small fixes to continue to support Python 2.7
(or earlier, though this hasn't been tested and isn't a goal).

The Job viewing, and its  use in tests, is removed, as this is
non-deterministic and perhaps not very useful outside of testing.
Simplifying Job is helpful since we hope to refactor it.