YAML input: accept "ranks:"
Split the function to process the file into separate functions to build
the Job and the Test.

Adjust the default resource setting logic for Job so that one need
not specify all three current defaults.

Have the Test inherit the name from the Job (which means that it gets
an underscore, if specified with a space, since Job enforces that, but
Test currently doesn't.)