Job: remove createJobOrdering
This is an incremental part of refactoring Job.

We plan for Tasks (sub-Jobs, currently) to be anonymous, so we
remove the function which uses their names to build output files,
instead using files with the name of the main Job.

These filenames will also likely change soon, as we move to a single
output file per Job, so don't put much effort into giving them good names.

Condense the getJobList() function as well.