Job, JobSequence: refactor into Job, Task
Introduce Task, which encapsulate most of what the base Job
class formerly defined, except for a name.

A Job is now something responsible for defining an ordered set of Tasks. This currently a linear ordering.

The Job __init__ function is changed to accept a Task or a list of Tasks. This might be too clunky, in which case it might be a good idea to re-introduce an option to create a single-Task Job directly.

The ExitCodeVerifier is significantly simplified here, and now produces output in its report.

An additional refactoring required is that currently, successful error
codes are attached to Task, whereas they should be attach to
ExitCodeVerifier (since only in the context of a Test can you say
whether a given error code is desireable, for a given task).

Some unimplemented stubs for specifying ranks/node were removed from the Launcher.