Add custom YAML subset parser
Add a custom parser for a subset of YAML files, sufficient to load our input files without the need for a non-standard module.

Only block collections are supported, not flow style.

Our logic is of course expected to be slower and less general than that in a proper module, but the argument for using it is that a primary selling point for SciATH is that you can just run it on your cluster with the system Python (if it's 2.7 or later). Our input files are small and simple enough that hopefully this custom parser won't be a maintenance burden.

SciATH is now free of heavy dependencies!

A single test of the parser is added.

See strictyaml (https://github.com/crdoconnor/strictyaml)
for a package that does things even better than PyYAML, for our
purposes, but which unfortunately would still introduce the dependencies
that we're trying to avoid.