Test, Harness: add test groups
Maintain a set of groups (strings with no spaces) for each Test.

Add new command line options to exclude tests which are in a list
of groups, and/or which are not in another list of tests.

As with the test subset options, these are specified as comma-separated
lists on the command line.