Line Verifier: add absolute tolerance for key-and-floats procedure
Allow for an absolute tolerance when using the LineVerifier with the included logic to compare lines starting with keys.

One or both tolerances may be specified. If either tolerance exists and passes, success is declared.

The output reporting is updated a little bit (but could still use a lot more work to be readable).

Reading form a YAML file is updated to allow an "atol:" key.
A new test is added, demonstrating this usage.