Use sentinel files to determine if a Job has launched or completed
Create the first sentinel file as soon as the command to launch has been executed (and in the case of running, locally, completed, since it blocks).

Create the second sentinel file once the command to execute have completed, either locally or in the batch file.

Only run the Verifier for a given Test once the underlying Job has been determined to have
been successfully run (as far as the Launcher is concerned).

This exposes the fact that we have a redundant set of status definitions in _teststatus.py, which are partially removed here, and will be completely removed in a future commit.