Verifier: return true/false passing value
Verifiers must always return a true or false value for whether a test
passed. This implies that when the verifier is invoked, the underlying
Job should have been successfully completed, as far as the Launcher can
tell (and this is now possible by checking a sentinel file).

Adjust the output of the execute() function for the Verifier class.

Remove a now-unneeded extra set of status definitions, as Verifiers
now return a true/false value, plus a short info string and a longer
report, and the larger set of statuses is only needed in _TestRunStatus.