Launcher: return information when a job cannot be launched
By having submit_job() return information, more elegantly deal
with the case where a test cannot be run due to lack of resources.
The Harness now uses this information to populate test run status
information, hopefully making it more clear what is going on when
doing something like trying to run an MPI test without an MPI-capable
Launcher.

Note that with test groups, the intended usage is that one should
explicitly exclude tests one does not actually want to run, hence global failure
is declared when tests are skipped for lack of resources.