YAML file parsing: add mechanism to capture environment variables
Generalize the find-replace mechanism to replace multiple values,
and allow for a user-provided list of environment variables (required)
to substitue values for, when prepended with "$" in the input YAML file.

This provides a way to pass parameters to the otherwise-static YAML
input files. This could get confusing so is considered advanced usage.

It uses a naive replacement scheme. A better one might be as is done
here: https://gitlab.com/petsc/petsc/-/blob/master/src/docs/sphinx_docs/ext/html5_petsc.py#L203