Launcher: Generate launch scripts from templates
Add an experimental feature where batch submission scripts are
generated by replacing special values in templates, as opposed
to being created by special functions in launcher.py.

This allows users to provide the necessary tweaks to their job submission scripts,
as is very common on clusters.

Templates are provided for for local, SLURM and LSF systems.
The templates are processed in a fairly naive way, which in its current
state is likely too brittle for general users.

The most brittle part of the templates may be that each is required
to have a set of three lines corresponding to setting the number of
threads, running with MPI, and running the actual command (and recording
its error code). If thread and/or MPI resources are not defined,
the script should run without those lines. For multiple Tasks, these
lines are repeated.

Activate by providing one of the templates in config_examples to the
new --template argument .

This implementation is incomplete in the sense that the templates do
not respect most of the information in the .conf file, including the
queue, account name, partition, and MPI launch command. Crucially,
however, the templates to do not yet contain any information about how
they are to be launched, so a compatible .conf file must be used
with a given template.

As suggested by the fact that Launcher() now accepts both a conf file
and a template file as input, a possible next step is to include *all*
Launcher configuration in the template file (possibly still generating
it, if desired, interactively from a meta-template for supported batch
systems).