Launcher: add local blocking "batch" system running shell scripts
As a step towards uniformly running all tests by passing a local
script (or batch file) to a submission system, add a new "local"
batch system which creates a shell script and runs it by passing to
the sh command. Note that this still blocks, but as with other
current batch systems, does not automatically verify.