Launcher: add launcher-level replacement for batch file templates
Add account and queue fields to the template processing.

Since these fields often have default values that the user may want to
use, add logic to replace the field if the value is defined in the
configuration, or delete the entire containing line if not defined. The
replacement string is named to try to alert to this to avoid unexpected
behavior.