YAML file parsing: update comment behavior
Do not strip after a "#" on all lines, as before.

Discourage the user from using trailing comments on data lines,
as we don't include this notion of comment in a "command", and  the
naive previous approach of stripping everything after any "#" is wrong,
e.g. POSIX shell prefix patterns use "#", and one may have a quoted "#"
in a legitimate command.