File input: read per-task ranks
An important change here is that the default for an unspecified number
of ranks is now 0, meaning that 1 rank can now be interpreted to mean
a job that requires MPI but runs on a single rank.

Currently, this can actually run e.g. "mpiexec -n 0", which is probably
not desired. A future change will eliminate the entire MPI launch line
for this case.