Harness, Launcher: move stdout output to Harness from Launcher
This enforces the intended behavior that only the Harness produces
output to stdout.

The is acheived by adding functions to the launcher to assemble the
command that will be used by submit_job() before it is called (required
since it may block). The harness then uses this function to print
the command that the Launcher will use.

Perhaps undesirable behavior: if a job is skipped
due to lack of MPI, a command is still printed, even though the
script named is never actually produced.